refactor: [#116] spec 경로 수정 (2)
